Utilise le composant avec une liste statique pour les mots clés contrôlés
