+++ front/gatsby/src/components/Write/metadata/yaml.js
@@ -1,20 +1,20 @@
 import yaml from 'js-yaml'
 
-export function cleanOutput (object){
+export function cleanOutput (object) {
   let cleaning = JSON.parse(JSON.stringify(object))
 
   if (ObjectIsEmpty(cleaning)) {
     return ''
   }
 
-  for (var propName in cleaning) { 
+  for (var propName in cleaning) {
     if (cleaning[propName] === null || cleaning[propName] === undefined || cleaning[propName] === "") {
       delete cleaning[propName];
     }
-    if(Array.isArray(cleaning[propName]) && cleaning[propName].length === 0){
+    if (Array.isArray(cleaning[propName]) && cleaning[propName].length === 0) {
       delete cleaning[propName];
     }
-    if(ObjectIsEmpty(cleaning[propName])) {
+    if (ObjectIsEmpty(cleaning[propName])) {
       delete cleaning[propName];
     }
   }
