--- front/gatsby/src/components/Write/yamleditor/default/rubriques.json
@@ -1,8 +1,0 @@
-[
-  {"label":"Essai","selected":false},
-  {"label":"Création","selected":false},
-  {"label":"Lecture","selected":false},
-  {"label":"Sommaire dossier","selected":false},
-  {"label":"Entretien","selected":false},
-  {"label":"Chronique","selected":false}
-]
